Move NOMIS model classes into `dto` sub-package because they too are “DTOs”, simply from another realm
