+++ src/main/kotlin/uk/gov/justice/digital/hmpps/incidentreporting/dto/nomis/NomisIncidentReport.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.incidentreporting.model.nomis
+package uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import io.swagger.v3.oas.annotations.media.Schema
