+++ src/test/kotlin/uk/gov/justice/digital/hmpps/incidentreporting/resource/MigrateResourceIntTest.kt
@@ -14,19 +14,19 @@ import uk.gov.justice.digital.hmpps.incidentreporting.integration.SqsIntegration
 import uk.gov.justice.digital.hmpps.incidentreporting.jpa.Report
 import uk.gov.justice.digital.hmpps.incidentreporting.jpa.repository.EventRepository
 import uk.gov.justice.digital.hmpps.incidentreporting.jpa.repository.ReportRepository
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.CodeDescription
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.History
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.HistoryQuestion
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.HistoryResponse
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.NomisIncidentReport
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.NomisIncidentStatus
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Offender
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.OffenderParty
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Question
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Requirement
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Response
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Staff
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.StaffParty
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.CodeDescription
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.History
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.HistoryQuestion
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.HistoryResponse
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.NomisIncidentReport
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.NomisIncidentStatus
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Offender
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.OffenderParty
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Question
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Requirement
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Response
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Staff
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.StaffParty
 import uk.gov.justice.digital.hmpps.incidentreporting.service.InformationSource
 import java.time.Clock
 import java.time.LocalDate
