+++ src/test/kotlin/uk/gov/justice/digital/hmpps/incidentreporting/service/SyncServiceTest.kt
@@ -19,13 +19,13 @@ import uk.gov.justice.digital.hmpps.incidentreporting.jpa.Report
 import uk.gov.justice.digital.hmpps.incidentreporting.jpa.Status
 import uk.gov.justice.digital.hmpps.incidentreporting.jpa.Type
 import uk.gov.justice.digital.hmpps.incidentreporting.jpa.repository.ReportRepository
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.CodeDescription
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.NomisIncidentReport
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.NomisIncidentStatus
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Offender
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.OffenderParty
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.Staff
-import uk.gov.justice.digital.hmpps.incidentreporting.model.nomis.StaffParty
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.CodeDescription
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.NomisIncidentReport
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.NomisIncidentStatus
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Offender
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.OffenderParty
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.Staff
+import uk.gov.justice.digital.hmpps.incidentreporting.dto.nomis.StaffParty
 import uk.gov.justice.digital.hmpps.incidentreporting.resource.NomisSyncRequest
 import uk.gov.justice.digital.hmpps.incidentreporting.resource.ReportNotFoundException
 import java.time.LocalDateTime
